+++ app/concerns/new/page.tsx
@@ -1,62 +1,5 @@
-'use client'
+import ConcernForm from "../_components/ConcernForm";
 
-import { ErrorMessage, Spinner } from '@/app/components';
-import { createConcernSchema } from "@/app/validationSchemas";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { Button, Callout, TextField } from "@radix-ui/themes";
-import axios from "axios";
-import "easymde/dist/easymde.min.css";
-import dynamic from "next/dynamic";
-import { useRouter } from "next/navigation";
-import { useState } from "react";
-import { Controller, useForm } from "react-hook-form";
-import { z } from "zod";
-
-const SimpleMDE = dynamic(
-  () => import('react-simplemde-editor'),
-  { ssr: false }
-)
-
-type ConcernForm = z.infer<typeof createConcernSchema>;
-
-export default function NewConcern() {
-  const router = useRouter();
-  const { register, control, handleSubmit, formState: { errors } } = useForm<ConcernForm>({
-    resolver: zodResolver(createConcernSchema)
-  });
-  const [error, setError] = useState('');
-  const [isSubmitting, setSubmitting] = useState(false);
-  const onSubmit = handleSubmit(async (data) => {
-    try {
-      setSubmitting(true);
-      await axios.post('/api/concerns', data);
-      router.push('/concerns');
-    } catch (error) {
-      setSubmitting(false);
-      setError('An unexpected error occurred.');
-    }
-  })
-
-  return (
-    <div className='max-w-xl'>
-      {error && <Callout.Root color='red' className='mb-5'>
-        <Callout.Text>{error}</Callout.Text>
-      </Callout.Root>}
-      <form
-        className='space-y-3'
-        onSubmit={onSubmit}>
-        <TextField.Root>
-          <TextField.Input placeholder='Title' {...register('title')} />
-        </TextField.Root>
-        <ErrorMessage>{errors.title?.message}</ErrorMessage>
-        <Controller
-          name='description'
-          control={control}
-          render={({ field }) => <SimpleMDE placeholder='Description' {...field} />}
-        />
-        <ErrorMessage>{errors.description?.message}</ErrorMessage>
-        <Button disabled={isSubmitting}>Submit New Concern {isSubmitting && <Spinner />}</Button>
-      </form>
-    </div>
-  )
+export default function NewConcernPage() {
+  return <ConcernForm />
 }
